switch to the 'libcoin' C API
Again using the example in r1773, determining the approximate null
distribution for maximum tests is now as fast as it was prior to
r1773.  For quadratic tests, it is now about 1.1x faster (35.6 vs 32.3
sec), but has higher memory consumption (1.14 vs 3.23 GB) due to
computing quadratic forms using R code instead of the optimized C code
called by 'libcoin::doTest()' (but that requires an object of class
"LinStatExpCov").

git-svn-id: svn://svn.r-forge.r-project.org/svnroot/coin@1784 edb9625f-4e0d-4859-8d74-9fd3b1da38cb